+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.4.0] - 2021-03-23
+## [0.4.0] - 2021-03-25
 
 ### New